Improve rerender dispatcher detection
