+++ package.json
@@ -14,13 +14,14 @@
 		"lint": "eslint 'packages/**/*.{ts,tsx,js,jsx}'",
 		"test": "pnpm test:karma && pnpm test:mocha",
 		"test:minify": "pnpm test:karma:minify",
-		"test:prod": "pnpm test:karma:prod",
+		"test:prod": "pnpm test:karma:prod && pnpm test:mocha:prod",
 		"test:karma": "cross-env COVERAGE=true karma start karma.conf.js --single-run",
 		"test:karma:minify": "cross-env COVERAGE=true MINIFY=true karma start karma.conf.js --single-run",
 		"test:karma:watch": "karma start karma.conf.js --no-single-run",
 		"test:karma:prod": "cross-env MINIFY=true NODE_ENV=production karma start karma.conf.js --single-run",
 		"test:karma:prod:watch": "cross-env NODE_ENV=production karma start karma.conf.js --no-single-run",
 		"test:mocha": "cross-env COVERAGE=true mocha --require packages/react/test/node/setup.js --recursive packages/react/test/node/**.test.tsx",
+		"test:mocha:prod": "cross-env COVERAGE=true NODE_ENV=production mocha --require packages/react/test/node/setup.js --recursive packages/react/test/node/**.test.tsx",
 		"docs:start": "cd docs && pnpm start",
 		"docs:build": "cd docs && pnpm build",
 		"docs:preview": "cd docs && pnpm preview",
